fix: add test and mv func location
