+++ src/commands/integration/deploy.ts
@@ -9,29 +9,10 @@ import {
   Webhook,
 } from '../../types';
 import integrationRpc from '../../rpc/integration';
-import { CompileResult, compileSol } from '../../compiler';
+import { CompileResult, compileSol, getLatestEvmVersion } from '../../compiler';
 import { Status } from '../../types/Integration';
 
 const defaultSpecFile = './henesis.yaml';
-
-function getEvmVersion(compilerVersion: string): string {
-  let evmVersion = 'petersburg';
-  const rawVersions: string[] = compilerVersion.split('.');
-
-  const versions: number[] = [
-    parseInt(rawVersions[0]),
-    parseInt(rawVersions[1]),
-    parseInt(rawVersions[2].split('-')[0]),
-  ];
-
-  if (versions[1] <= 5) {
-    if (versions[2] < 5) {
-      evmVersion = 'byzantium';
-    }
-  }
-
-  return evmVersion;
-}
 
 async function getAbi(
   path: string,
@@ -40,7 +21,7 @@ async function getAbi(
 ): Promise<any> {
   const result: CompileResult = await compileSol(path, {
     solcVersion: compilerVersion,
-    evmVersion: getEvmVersion(compilerVersion),
+    evmVersion: getLatestEvmVersion(compilerVersion),
   });
 
   return result.getAbi(contractName);
